Fix: liff.initを修正 #48
