+++ frontend/src/components/pages/Calendar.tsx
@@ -1,7 +1,5 @@
-import { memo, useCallback, useEffect, useState, VFC } from "react"
-import axios from "axios"
+import { memo, useCallback, useState, VFC } from "react"
 import dayjs from "dayjs"
-import liff from "@line/liff/dist/lib"
 import { useDisclosure } from "@chakra-ui/react"
 
 import FullCalendar, { EventClickArg } from "@fullcalendar/react"
@@ -10,12 +8,14 @@ import interactionPlugin, { DateClickArg } from "@fullcalendar/interaction"
 import allLocales from '@fullcalendar/core/locales-all'
 
 import { Food } from "../../type/Food"
+import { useLogin } from "../../hooks/useLogin"
 import { useFoods } from "../../hooks/useFoods"
 import { ClickDayFoodsDetailDrawer } from "../organisms/ClickDayFoodsDetailDrawer"
 
 export const Calendar: VFC = memo(() => {
   const [ clickDayFoods, setClickDayFoods ] = useState<Array<Food>>()
   const [ clickDay, setClickDay ] = useState("")
+  const { login } = useLogin()
   const { foods, setFoods } = useFoods()
   const { isOpen, onClose, onOpen } = useDisclosure()
 
@@ -36,27 +36,7 @@ export const Calendar: VFC = memo(() => {
   }, [onOpen, foods])
 
   // ログインする
-  useEffect(() => {
-    liff.init({ liffId: process.env.REACT_APP_LIFF_ID_CALENDAR as string })
-        .then(() => {
-          // ログインしていなかったらログインする
-          if (!liff.isLoggedIn()) {
-            liff.login()
-          }
-
-          const idToken = liff.getIDToken()
-          const params = new URLSearchParams()
-          params.append("idToken", idToken as string)
-          axios.post("/users", params)
-               .then(res => {
-                // IdTokenの有効期限が切れたらログアウトする
-                if (res.data.error_description === 'IdToken expired.') {
-                  liff.logout()
-                }
-               })
-               .catch(e => console.error(e))
-        })
-  }, [])
+  login({ liffId: process.env.REACT_APP_LIFF_ID_CALENDAR as string })
 
   return (
     <>
